// client/src/app/models/Project.ts
interface Project {
  donated: string,
  title: string,
  sponsors: string[],
  sponsorsCount: string,
  finished: boolean,
  created_at: string,
  updated_at: string,
  location: string
  description: string
  initiator: string
  amount: string
  started_at: string
  finished_at: string
  img: string
  user: string
  _id: string
}
export default Project;